requirements-dev.txt
# Core Dependencies
networkx>=3.0.0
openai>=1.0.0
orjson>=3.9.0
pathspec
pydantic>=2.0.0
tree-sitter>=0.20.0
tree-sitter-language-pack
tenacity

# TDD/QA Tools
pytest
pytest-cov
pytest-sugar
pytest-asyncio
python-dotenv
ruff
mypy
types-networkx

# Documentation
mkdocs-material
mkdocstrings[python]
